+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 0.5.7 (unreleased)
 ------------------
 
-- Add upgrade to clean portal_skins from all cpskin (unused) skins
+- Add upgrade to clean portal_skins from all cpskin (unused) skins / folders
   [laulaz]
 
 - Fix icon-home commented code by mistake